DTPM module overhaul
Instead of parsing the smsbus website, get the SOAP API response directly and parse the data from there.
That allows for more flexible data gathering. It makes use of this by creating a custom dict for easier interpretation of the data.

This revision also fixes the issue that made the module return data from the second bus arriving to the bus stop instead of the first one.